Move parsing logic into parser.ts
